requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.1
pinecone==5.0.0   # or latest

PyPDF2==3.0.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
